fix: extract shared inputStyle, reportError, kebab-case step filenames, ts-expect-error Blob cast
