+++ src/features/leagues/components/create-league/form-styles.ts
@@ -1,0 +1,12 @@
+import { mflColors } from '../../../../constants/colors';
+
+export const inputStyle = {
+  backgroundColor: mflColors.white,
+  borderWidth: 1,
+  borderColor: mflColors.border,
+  borderRadius: 12,
+  paddingHorizontal: 16,
+  paddingVertical: 12,
+  fontSize: 15,
+  color: mflColors.text,
+} as const;
